requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
redis==5.0.1
openai==1.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6 